requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
autoflake==2.3.1
autoimport==1.4.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
idna==3.7
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
maison==1.4.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipdeptree==2.18.1
pipreqs==0.5.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
pure-eval==0.2.2
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
Pygments==2.17.2
pyodbc==5.1.0
pyprojroot==0.3.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.0.2
referencing==0.35.0
requests==2.31.0
rpds-py==0.18.0
sh==2.0.6
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
toml==0.10.2
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
xdg==6.0.0
yarg==0.1.9
maturin
underthesea
unidecode
pandas
python-dotenv
matplotlib